more warning cleanup
const, unused, mismatched types, field initializers

+++ Classes/OGVDecoderAV.m
@@ -311,7 +311,7 @@
         channelPtrs[i] = &floatPtr[i * samples];
     }
     OGVAudioBuffer *audioBuffer = [[OGVAudioBuffer alloc] initWithPCM:channelPtrs
-                                                              samples:samples
+                                                              samples:(unsigned int)samples
                                                                format:format
                                                             timestamp:time];
     free(channelPtrs);
